完成调度任务中心对接

+++ z-dtx-job/src/main/java/org/zipper/job/controller/RuleController.java
@@ -38,7 +38,7 @@ public class RuleController {
     }
 
     @GetMapping("find")
-    public ResponseEntity<Rule> findOne(@RequestParam Integer id) {
+    public ResponseEntity<Rule> findOne(@RequestParam Long id) {
         Rule rule = this.ruleService.queryOne(id);
         return ResponseEntity.success(rule);
     }
