+++ z-dtx-job/src/main/java/org/zipper/job/service/IRuleService.java
@@ -13,7 +13,7 @@ public interface IRuleService {
 
     PageInfo<RuleVO> listPage(RuleQueryParams params, int pageSize, int pageNum);
 
-    Rule queryOne(Integer id);
+    Rule queryOne(Long id);
 
     boolean updateOne(RuleDTO dto);
